Make some fixes.
- Make some functions const.
- Added CursorPassThroughEnabled backward compatibility.
- Added CEGUI/Config for FreeType and PCRERegexMatcher.
- Fixed warning about uint missing.

+++ CEGUINullRenderer/CEGUI/RendererModules/Null/GeometryBuffer.cpp
@@ -130,7 +130,7 @@ Texture* NullGeometryBuffer::getActiveTexture() const
 //----------------------------------------------------------------------------//
 uint NullGeometryBuffer::getVertexCount() const
 {
-    return d_vertices.size();
+    return (uint)d_vertices.size();
 }
 
 //----------------------------------------------------------------------------//
